move shared code to shared package
