+++ src/main/scala/akka/persistence/dynamodb/journal/DynamoDBHelper.scala
@@ -3,20 +3,22 @@
  */
 package akka.persistence.dynamodb.journal
 
-import com.amazonaws.{ AmazonServiceException, AmazonWebServiceRequest }
+import com.amazonaws.{AmazonServiceException, AmazonWebServiceRequest}
 import com.amazonaws.handlers.AsyncHandler
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBAsyncClient
 import com.amazonaws.services.dynamodbv2.model._
 import akka.actor.Scheduler
 import akka.event.LoggingAdapter
 import akka.pattern.after
-import java.util.{ concurrent => juc }
+import java.util.{concurrent => juc}
+
 import scala.collection.JavaConverters._
-import scala.concurrent.{ ExecutionContext, Future, Promise }
+import scala.concurrent.{ExecutionContext, Future, Promise}
 import scala.concurrent.duration._
 import scala.reflect.ClassTag
 import scala.util.control.NoStackTrace
 import akka.actor.ActorRef
+import akka.persistence.dynamodb.DynamoDBConfig
 
 case class LatencyReport(nanos: Long, retries: Int)
 private class RetryStateHolder(var retries: Int = 10, var backoff: FiniteDuration = 1.millis)
